ref: move helpers to functions/mediaHelpers
